Entity pages and ECS Core

+++ GameProject/Engine/Utils/Logger.hpp
@@ -66,7 +66,7 @@ template<typename ... Args>
 std::string Logger::formatString(const std::string& format, Args&& ... args)
 {
     // Create a char buffer to hold the formatted string and a null termination
-    size_t size = snprintf(nullptr, 0, format.c_str(), std::forward<Args>(args) ...) + 1;
+    size_t size = (size_t)snprintf(nullptr, 0, format.c_str(), std::forward<Args>(args) ...) + 1;
     std::unique_ptr<char[]> buffer(new char[size]);
 
     // Format string
